+++ src/app/cart/components/confirm-order/confirm-order.component.ts
@@ -1,17 +1,21 @@
-import { CartService } from './../../services/cart.service';
-import { Observable } from 'rxjs';
+import { Component, OnInit, OnDestroy } from '@angular/core';
+import { ActivatedRoute, Params } from '@angular/router';
+import { Observable, from, Subscription } from 'rxjs';
+import { map, mergeMap, tap } from 'rxjs/operators';
+
 import { IUniqProduct } from '@models/product.model';
 import { ProductService } from './../../../products/services/product.service';
-import { ActivatedRoute, Params } from '@angular/router';
-import { Component, OnInit } from '@angular/core';
+import { CartService } from './../../services/cart.service';
 
 @Component({
   selector: 'app-confirm-order',
   templateUrl: './confirm-order.component.html',
   styleUrls: ['./confirm-order.component.scss']
 })
-export class ConfirmOrderComponent implements OnInit {
+export class ConfirmOrderComponent implements OnInit, OnDestroy {
   products$: Observable<IUniqProduct[]>;
+  subscriptions = new Subscription();
+
   constructor(
     private route: ActivatedRoute,
     private productService: ProductService,
@@ -19,8 +23,16 @@ export class ConfirmOrderComponent implements OnInit {
   ) {}
 
   ngOnInit() {
-    this.route.queryParams.subscribe(this.getProducts());
+    const subscription = this.route.queryParams.subscribe(params =>
+      this.getProducts(params)
+    );
     this.products$ = this.cartService.products$;
+
+    this.subscriptions.add(subscription);
+  }
+
+  ngOnDestroy() {
+    this.subscriptions.unsubscribe();
   }
 
   onProductChange(product: IUniqProduct) {
@@ -37,24 +49,28 @@ export class ConfirmOrderComponent implements OnInit {
     );
   }
 
-  private getProducts(): (value: Params) => void {
-    return params => {
-      Object.keys(params)
-        .map(id => this.productService.getProductById(id))
-        .forEach(promise =>
-          promise.then(product => {
-            const uniqProduct = {
-              ...product,
-              quantity: params[product.id],
-              product
-            };
+  private getProducts(params: Params): void {
+    const subscription = from(Object.keys(params))
+      .pipe(
+        mergeMap(id => this.productService.getProductById(id)),
+        map(product => {
+          const uniqProduct = {
+            ...product,
+            quantity: params[product.id],
+            product
+          };
 
-            this.cartService.addProduct(
-              uniqProduct,
-              parseInt(uniqProduct.quantity)
-            );
-          })
-        );
-    };
+          return uniqProduct;
+        }),
+        tap(uniqProduct =>
+          this.cartService.addProduct(
+            uniqProduct,
+            parseInt(uniqProduct.quantity)
+          )
+        )
+      )
+      .subscribe();
+
+    this.subscriptions.add(subscription);
   }
 }